backend/app/ml/requirements.txt
# AI/ML Dependencies
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
librosa==0.10.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Blockchain Dependencies
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# Image Processing
Pillow==10.1.0
imageio==2.33.0
scipy==1.11.4

# Video Processing
moviepy==1.0.3
av==11.0.0

# Audio Processing
soundfile==0.12.1
pydub==0.25.1

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
cryptography==41.0.7
hashlib-additional==1.1.0
